refactor: Name, Position의 테스트 코드 추가
