+++ src/test/java/racingcar/RacingGameTest.java
@@ -10,9 +10,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import racingcar.domain.Car;
+import racingcar.domain.car.Car;
 import racingcar.domain.Cars;
-import racingcar.domain.Name;
+import racingcar.domain.car.Name;
 import racingcar.domain.RacingGame;
 import racingcar.domain.Trial;
